Changed how we got line 45 - just did a count. The text that was searched for before, appeared multiple times in document. It just skips first 44 lines, and then parses 45. Trimmed whitespace from line before casting to float.
Added check to make sure line 45 has data (length > 0) - this may not be necessary, but is a sanity check for running code while txt files are still being populated.

Added code to be able to run this even if all files aren't in directory, if file doesn't open it just continues and checks next file. This will allow us to run this on multiple machines and we don't need to copy all files into one directory/machine.